Address warnings from Windows CI job in composable player and recorder

+++ rosbag2_transport/src/rosbag2_transport/config_options_from_node_params.cpp
@@ -124,7 +124,7 @@ PlayOptions get_play_options_from_node_params(rclcpp::Node & node)
   auto desc_rate = param_utils::float_param_description(
     "Playback rate (hz)",
     0.000001,
-    std::numeric_limits<float>::max());
+    static_cast<float>(std::numeric_limits<float>::max()));
   play_options.rate =
     static_cast<float>(node.declare_parameter<float>("play.rate", 1.0, desc_rate));
